修改数据类型

+++ src/drivers/linux_ina219/linux_ina219.cpp
@@ -59,8 +59,8 @@ void linux_ina219::usage() {
 //---------------------------------------------------------------------------------------------//
 void linux_ina219::running(int argc,char**argv) {
 	PX4_INFO("持续测量");
-	float v;
-	float sv;
+	double v;
+	double sv;
 	//float a;
 	//a = linux_ina219::ina219->getCurrentMa();
 	//a = 1.0;
